+++ apm-commons/apm-util/src/main/java/org/apache/skywalking/apm/util/ConfigInitializer.java
@@ -25,6 +25,7 @@ import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
@@ -32,6 +33,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Properties;
 import java.util.Set;
+import java.util.TreeMap;
 import java.util.TreeSet;
 
 /**
@@ -39,6 +41,7 @@ import java.util.TreeSet;
  * <p>
  */
 public class ConfigInitializer {
+
     public static void initialize(Properties properties, Class<?> rootConfigType) throws IllegalAccessException {
         initNextLevel(properties, rootConfigType, new ConfigDesc());
     }
@@ -49,48 +52,50 @@ public class ConfigInitializer {
             if (Modifier.isPublic(field.getModifiers()) && Modifier.isStatic(field.getModifiers())) {
                 String configKey = (parentDesc + "." + field.getName()).toLowerCase();
                 Class<?> type = field.getType();
-
-                if (type.equals(Map.class)) {
+                if (Map.class.isAssignableFrom(type)) {
                     /*
-                     * Map config format is, config_key[map_key]=map_value
-                     * Such as plugin.opgroup.resttemplate.rule[abc]=/url/path
+                     * Map config format is, config_key[map_key]=map_value, such as plugin.opgroup.resttemplate.rule[abc]=/url/path
+                     * "config_key[]=" will generate an empty Map , user could use this mechanism to set an empty Map
                      */
-                    // Deduct two generic types of the map
                     ParameterizedType genericType = (ParameterizedType) field.getGenericType();
                     Type[] argumentTypes = genericType.getActualTypeArguments();
-
-                    Type keyType = null;
-                    Type valueType = null;
-                    if (argumentTypes != null && argumentTypes.length == 2) {
-                        // Get key type and value type of the map
-                        keyType = argumentTypes[0];
-                        valueType = argumentTypes[1];
-                    }
-                    Map map = (Map) field.get(null);
-                    // Set the map from config key and properties
-                    setForMapType(configKey, map, properties, keyType, valueType);
-                } else if (Collection.class.isAssignableFrom(type)) {
-                    ParameterizedType genericType = (ParameterizedType) field.getGenericType();
-                    Type argumentType = genericType.getActualTypeArguments()[0];
-                    Collection collection = convertToCollection(argumentType, type, properties.getProperty(configKey));
-                    field.set(null, collection);
-                } else {
-                    /*
-                     * Others typical field type
-                     */
-                    String value = properties.getProperty(configKey);
-                    // Convert the value into real type
-                    final Length lengthDefine = field.getAnnotation(Length.class);
-                    if (lengthDefine != null) {
-                        if (value != null && value.length() > lengthDefine.value()) {
-                            value = value.substring(0, lengthDefine.value());
+                    Type keyType = argumentTypes[0];
+                    Type valueType = argumentTypes[1];
+                    // A chance to set an empty map
+                    if (properties.containsKey(configKey + "[]")) {
+                        Map currentValue = (Map) field.get(null);
+                        if (currentValue != null && !currentValue.isEmpty()) {
+                            field.set(null, initEmptyMap(type));
+                        }
+                    } else {
+                        // Set the map from config key and properties
+                        Map map = readMapType(type, configKey, properties, keyType, valueType);
+                        if (map.size() != 0) {
+                            field.set(null, map);
                         }
                     }
-                    Object convertedValue = convertToTypicalType(type, value);
-                    if (convertedValue != null) {
-                        field.set(null, convertedValue);
+                } else if (properties.containsKey(configKey)) {
+                    //In order to guarantee the default value could be reset as empty , we parse the value even if it's blank
+                    String propertyValue = properties.getProperty(configKey, "");
+                    if (Collection.class.isAssignableFrom(type)) {
+                        ParameterizedType genericType = (ParameterizedType) field.getGenericType();
+                        Type argumentType = genericType.getActualTypeArguments()[0];
+                        Collection collection = convertToCollection(argumentType, type, propertyValue);
+                        field.set(null, collection);
+                    } else {
+                        // Convert the value into real type
+                        final Length lengthDefine = field.getAnnotation(Length.class);
+                        if (lengthDefine != null && propertyValue.length() > lengthDefine.value()) {
+                            StringUtil.cut(propertyValue, lengthDefine.value());
+                            System.err.printf("The config value will be truncated , because the length max than %d : %s -> %s%n", lengthDefine.value(), configKey, propertyValue);
+                        }
+                        Object convertedValue = convertToTypicalType(type, propertyValue);
+                        if (convertedValue != null) {
+                            field.set(null, convertedValue);
+                        }
                     }
                 }
+
             }
         }
         for (Class<?> innerConfiguration : recentConfigType.getClasses()) {
@@ -131,10 +136,9 @@ public class ConfigInitializer {
      * @return converted value or null
      */
     private static Object convertToTypicalType(Type type, String value) {
-        if (value == null || type == null) {
+        if (StringUtil.isBlank(value)) {
             return null;
         }
-
         Object result = null;
         if (String.class.equals(type)) {
             result = value;
@@ -160,21 +164,23 @@ public class ConfigInitializer {
     /**
      * Set map items.
      *
+     * @param type       the filed type
      * @param configKey  config key must not be null
-     * @param map        map to set must not be null
      * @param properties properties must not be null
      * @param keyType    key type of the map
      * @param valueType  value type of the map
      */
-    private static void setForMapType(String configKey, Map<Object, Object> map, Properties properties,
-                                      final Type keyType, final Type valueType) {
-        Objects.requireNonNull(configKey);
-        Objects.requireNonNull(map);
-        Objects.requireNonNull(properties);
+    private static Map readMapType(Class<?> type,
+                                   String configKey,
+                                   Properties properties,
+                                   final Type keyType,
+                                   final Type valueType) {
 
+        Objects.requireNonNull(configKey);
+        Objects.requireNonNull(properties);
+        Map<Object, Object> map = initEmptyMap(type);
         String prefix = configKey + "[";
         String suffix = "]";
-
         properties.forEach((propertyKey, propertyValue) -> {
             String propertyStringKey = propertyKey.toString();
             if (propertyStringKey.startsWith(prefix) && propertyStringKey.endsWith(suffix)) {
@@ -196,8 +202,18 @@ public class ConfigInitializer {
                 map.put(keyObj, valueObj);
             }
         });
+        return map;
     }
 
+    private static Map<Object, Object> initEmptyMap(Class<?> type) {
+        if (type.equals(Map.class) || type.equals(HashMap.class)) {
+            return new HashMap<>();
+        } else if (type.equals(TreeMap.class)) {
+            return new TreeMap<>();
+        } else {
+            throw new UnsupportedOperationException("Config parameter type support Map,HashMap,TreeMap");
+        }
+    }
 }
 
 class ConfigDesc {